move eval_bc to its own module
